gobject: Fix various implicit conversions from size_t to smaller types
Basically various trivial instances of the following MSVC compiler
warning:
```
../gio/gio-tool-set.c(50): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
```

+++ gobject/gobject-query.c
@@ -47,7 +47,7 @@ show_nodes (GType        type,
 	    const gchar *indent)
 {
   GType   *children;
-  guint i;
+  size_t i;
   
   if (!type)
     return;
